moving helpers into the library
