+++ src/index.js
@@ -19,7 +19,6 @@ let store = createStore(
   )
 )
 
-
 ReactDOM.render(
   <Provider store={ store }>
     <App />
